+++ conda_environment.yml
@@ -1,4 +1,4 @@
-name: eofunctions
+name: openeo_processes
 channels:
 - defaults
 dependencies:
