+++ docs/conf.py
@@ -35,7 +35,7 @@ except ImportError:
     from sphinx import apidoc
 
 output_dir = os.path.join(__location__, "api")
-module_dir = os.path.join(__location__, "../src/eofunctions")
+module_dir = os.path.join(__location__, "../src/openeo_processes")
 try:
     shutil.rmtree(output_dir)
 except FileNotFoundError:
@@ -148,7 +148,7 @@ html_theme = 'alabaster'
 # The name for this set of Sphinx documents.  If None, it defaults to
 # "<project> v<release> documentation".
 try:
-    from eofunctions import __version__ as version
+    from openeo_processes import __version__ as version
 except ImportError:
     pass
 else:
@@ -213,7 +213,7 @@ html_static_path = ['_static']
 # html_file_suffix = None
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = 'eofunctions-doc'
+htmlhelp_basename = 'openeo_processes-doc'
 
 
 # -- Options for LaTeX output --------------------------------------------------
