+++ setup.cfg
@@ -3,10 +3,10 @@
 # http://setuptools.readthedocs.io/en/latest/setuptools.html#configuring-setup-using-setup-cfg-files
 
 [metadata]
-name = eoFunctions
+name = openeo_processes
 description = Add a short description here!
-author = Luca Foresta
-author-email = luca.foresta@eodc.eu
+author = Claudio Navacchi
+author-email = claudio.navacchi@geo.tuwien.ac.at
 license = mit
 url = http://...
 long-description = file: README.rst
@@ -41,7 +41,7 @@ exclude =
 
 [options.extras_require]
 # Add here additional requirements for extra features, to install with:
-# `pip install eoFunctions[PDF]` like:
+# `pip install openeo_processes[PDF]` like:
 # PDF = ReportLab; RXP
 
 [test]
@@ -54,7 +54,7 @@ addopts = tests
 # e.g. --cov-report html (or xml) for html/xml output or --junitxml junit.xml
 # in order to write a coverage file that can be read by Jenkins.
 addopts =
-    --cov eofunctions --cov-report term-missing
+    --cov openeo_processes --cov-report term-missing
     --verbose
 norecursedirs =
     dist
@@ -91,4 +91,4 @@ exclude =
 # PyScaffold's parameters when the project was created.
 # This will be used when updating. Do not change!
 version = 3.0.3
-package = eofunctions
+package = openeo_processes
