+++ src/openeo_processes/texts.py
