+++ tests/test_arrays.py
@@ -5,7 +5,7 @@ https://openeo.org/documentation/1.0/processes.html
 
 import unittest
 import numpy as np
-import eofunctions as eof
+import openeo_processes as eof
 
 
 class ArrayTester(unittest.TestCase):
